More thresholding:
* Have a range from 0 to 255
* Allow also to change the threshold of a mesh
* Only do it for a mesh if it was calculated from an ImagePlus
* If volume or orthoslice, change it interactively, if mesh, just
  ask for the number and wait for OK, since the mesh has to be recalculated.

+++ ij3d/Content.java
@@ -199,8 +199,8 @@ public abstract class Content extends BranchGroup {
 	public abstract void colorUpdated(Color3f oldColor, Color3f newColor);
 	public abstract void channelsUpdated(boolean[] channels);
 	public abstract void transparencyUpdated(float transparency);
-	public abstract void setThreshold(double t);
-	public abstract double getThreshold();
+	public abstract void setThreshold(int t);
+	public abstract int getThreshold();
 	public abstract void flush();
 }
